requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
autoflake==2.3.1
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
Flask-HTTPAuth==4.8.0
Flask-Session==0.8.0
frozenlist==1.5.0
greenlet==3.2.0
griffe==1.7.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
identify==2.6.9
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.6.0
msgspec==0.19.0
multidict==6.4.3
mypy-extensions==1.0.0
nodeenv==1.9.1
openai==1.73.0
openai-agents==0.0.9
packaging==24.2
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
pyflakes==3.3.2
PyJWT==2.10.1
pypdfium2==4.30.1
pytest==8.3.5
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.46.2
stytch==13.1.0
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
virtualenv==20.30.0
Werkzeug==3.1.3
yarl==1.19.0
